+++ packages/template-retail-react-app/app/page-designer/utils.js
@@ -5,6 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+// TODO - move this to a more general location
 /**
  * Determines whether the specified URL is absolute
  *